Basic Port to React.
So far this contains a basic port of the old sidebar to also work with
react.
By basic, I mean that the sidebar selection order is exactly the same
(as much as is possible right now).

In order to work with the tumblr css map, all functions related to
sidebar.add are now asynchronous.